Move loop_state to different file
