+++ .gitignore
@@ -11,8 +11,6 @@ __pycache__/
 *.swp
 tmp
 .zed
-.cargo/config.toml
-
 CLAUDE.md
 AGENTS.md
 FILE STRUCTURE.md
