+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
   "crates/lexer",
   "crates/parser",
   "crates/renderer",
+  "crates/runtime",
   "crates/stdlib",
   "crates/stdlib-macros",
   "crates/structs",
@@ -36,12 +37,13 @@ smallvec = "1.15.1"
 unicode-segmentation = "1.12.0"
 open = "5.3.2"
 dirs = "5.0.0"
-serde = "1.0.228"
+serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 schemars = "1.0.4"
 sha2 = "0.10.9"
 futures = "0.3.31"
 wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
 smol = "2.0.2"
 notify-debouncer-full = "0.7.0"
 itertools = "0.14.0"
@@ -77,9 +79,10 @@ lexer = { path = "crates/lexer" }
 libtess2 = { path = "crates/libtess2" }
 parser = { path = "crates/parser" }
 renderer = { path = "crates/renderer" }
+runtime = { path = "crates/runtime" }
 inventory = "0.3"
 stdlib-macros = { path = "crates/stdlib-macros" }
-stdlib = { path = "crates/stdlib" }
+stdlib = { path = "crates/stdlib", default-features = false }
 structs = { path = "crates/structs" }
 monocurl = { path = "crates/monocurl" }
 ui_cli_shared = { path = "crates/ui_cli_shared" }
