+++ crates/stdlib/src/mesh/graphs.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-#[cfg(not(target_arch = "wasm32"))]
+#[cfg(all(not(target_arch = "wasm32"), feature = "native-latex"))]
 use executor::executor::TextRenderQuality;
 use executor::{error::ExecutorError, executor::Executor, heap::with_heap, value::Value};
 use geo::{
@@ -87,7 +87,7 @@ fn ensure_surface_triangles(kind: &str, tris: usize) -> Result<(), ExecutorError
     ensure_limit(kind, tris, MAX_SURFACE_TRIANGLES)
 }
 
-#[cfg(not(target_arch = "wasm32"))]
+#[cfg(all(not(target_arch = "wasm32"), feature = "native-latex"))]
 fn text_render_quality(executor: &Executor) -> latex::RenderQuality {
     match executor.text_render_quality() {
         TextRenderQuality::Normal => latex::RenderQuality::Normal,
@@ -638,7 +638,7 @@ fn render_axis_tex_tree(
     scale: f32,
     name: &'static str,
 ) -> Result<Option<MeshTree>, ExecutorError> {
-    #[cfg(target_arch = "wasm32")]
+    #[cfg(any(target_arch = "wasm32", not(feature = "native-latex")))]
     {
         let _ = (executor, tex, scale);
         return Err(ExecutorError::invalid_invocation(format!(
@@ -646,7 +646,7 @@ fn render_axis_tex_tree(
         )));
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
+    #[cfg(all(not(target_arch = "wasm32"), feature = "native-latex"))]
     {
         let meshes = latex::render_tex_with_quality(tex, scale, text_render_quality(executor))
             .map_err(|error| {
