+++ crates/web_runtime/Cargo.toml
@@ -10,7 +10,9 @@ crate-type = ["cdylib", "rlib"]
 bytecode.workspace = true
 executor.workspace = true
 renderer.workspace = true
+runtime.workspace = true
 stdlib.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen.workspace = true
+wasm-bindgen-futures.workspace = true
