+++ crates/web_runtime/src/lib.rs
@@ -1,11 +1,9 @@
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
 
-pub const BYTECODE_ABI_VERSION: u32 = 0;
-
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 pub struct Runtime {
-    native_function_count: usize,
+    controller: runtime::RuntimeController,
 }
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
@@ -13,20 +11,68 @@ impl Runtime {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen(constructor))]
     pub fn new() -> Self {
         Self {
-            native_function_count: stdlib::registry::registry().len(),
+            controller: runtime::RuntimeController::new(),
         }
     }
 
-    pub fn bytecode_abi_version(&self) -> u32 {
-        BYTECODE_ABI_VERSION
+    pub fn monocurl_version(&self) -> String {
+        bytecode::MONOCURL_VERSION.to_string()
+    }
+
+    pub fn supports_monocurl_version(&self, version: &str) -> bool {
+        version == bytecode::MONOCURL_VERSION
     }
 
     pub fn native_function_count(&self) -> usize {
-        self.native_function_count
+        stdlib::registry::registry().len()
     }
 
     pub fn bytecode_instruction_size(&self) -> usize {
         std::mem::size_of::<bytecode::Instruction>()
+    }
+
+    pub fn needs_work(&self) -> bool {
+        self.controller.needs_work()
+    }
+
+    pub fn is_playing(&self) -> bool {
+        self.controller.is_playing()
+    }
+
+    pub fn seek_to(&self, slide: usize, time: f64) {
+        self.controller.apply_command(
+            runtime::RuntimeCommand::SeekTo {
+                target: executor::time::Timestamp::new(slide, time),
+            },
+            0.0,
+        );
+    }
+
+    pub fn toggle_play(&self, now_seconds: f64) {
+        self.controller
+            .apply_command(runtime::RuntimeCommand::TogglePlay, now_seconds);
+    }
+
+    pub fn set_presentation_mode(&self) {
+        self.controller.apply_command(
+            runtime::RuntimeCommand::SetPlaybackMode(runtime::PlaybackMode::Presentation),
+            0.0,
+        );
+    }
+
+    pub fn set_preview_mode(&self) {
+        self.controller.apply_command(
+            runtime::RuntimeCommand::SetPlaybackMode(runtime::PlaybackMode::Preview),
+            0.0,
+        );
+    }
+
+    pub async fn step(&self, now_seconds: f64) -> usize {
+        self.controller
+            .run_iteration(now_seconds)
+            .await
+            .snapshots
+            .len()
     }
 }
 
@@ -38,7 +84,12 @@ impl Default for Runtime {
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 pub fn engine_version() -> String {
-    env!("CARGO_PKG_VERSION").to_string()
+    monocurl_version()
+}
+
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
+pub fn monocurl_version() -> String {
+    bytecode::MONOCURL_VERSION.to_string()
 }
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
